requirements.txt
Flask==2.2.2
python-sane==2.9.1
Pillow==9.4.0
